prune: move logic into repository package
